+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Unique ability stats for Thunderbird
 * Unique ability stat for Frost
 * Unique ability stat for Caveira
+* DP27 weapon
 
 ### Changed
 * Unique ability for Tachanka